+++ src/main/java/gift/kakao/login/controller/KakaoLoginController.java
@@ -1,0 +1,24 @@
+package gift.kakao.login.controller;
+
+import gift.kakao.login.service.KakaoLoginService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class KakaoLoginController {
+    private final KakaoLoginService kakaoLoginService;
+    public KakaoLoginController(KakaoLoginService kakaoLoginService){
+        this.kakaoLoginService = kakaoLoginService;
+    }
+    @GetMapping("")
+    public ResponseEntity<String> handleOAuthCallback(
+        @RequestParam(value = "code") String code) {
+        System.out.println("token: " + code);
+        String accessToken = kakaoLoginService.getAccessToken(code);
+        System.out.println("access_token: " + accessToken);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+}
