+++ src/main/java/gift/kakao/login/dto/KakaoTokenResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.login.dto;
+package gift.kakao.login.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
